refactor(runtime): split session_auth.rs god object into modules
