+++ scripts/.god-object-baseline.txt
@@ -1,7 +1,6 @@
 # God-object baseline (LOC ceilings). Files here may only SHRINK.
 # Regenerate after a real split: scripts/check-file-sizes.sh --update
 # Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
-1065 sandbox-runtime/src/session_auth.rs
 738 ui/src/routes/sandboxes.$id.tsx
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
 634 ui/src/routes/instances.$id.tsx
